+++ build.gradle
@@ -13,7 +13,7 @@ apply plugin: 'net.minecraftforge.gradle'
 apply plugin: 'eclipse'
 apply plugin: 'maven-publish'
 
-version = '2.2'
+version = '2.3'
 group = 'thatguy.mod.miningspeed2' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = 'miningspeed2'
