+++ src/main/resources/META-INF/mods.toml
@@ -19,7 +19,7 @@ modId="miningspeed2" #mandatory
 # The version number of the mod - there's a few well known ${} variables useable here or just hardcode it
 # ${file.jarVersion} will substitute the value of the Implementation-Version as read from the mod's JAR file metadata
 # see the associated build.gradle script for how to populate this completely automatically during a build
-version="2.2" #mandatory
+version="2.3" #mandatory
  # A display name for the mod
 displayName="Precision Mining" #mandatory
 # A URL to query for updates for this mod. See the JSON update specification https://mcforge.readthedocs.io/en/latest/gettingstarted/autoupdate/
